pyproject.toml
[project]
name = "watersortbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "scikit-learn>=1.7.2",
]
